feat: stores only 1 read in mem if in dedup mode
before, we would store all reads for a given UMI regardless of whether
we were deduplicating or just grouping, leading to large memory usage.

Now, if deduplicating, we just retain the highest quality read for each
unique sequence, for each UMI. This saves lots of memory when we have
lots of reads with the same UMI, and the same sequence.